scanner/requirements.txt
mariadb==1.0.11
python_masscan==0.1.6
tqdm==4.64.1
